// src/app/models/CrearUsuarioModel/CrearUsuario.model.ts
export class CrearUsuario 
{
    nombreCOmpleto: string;
    usuario: string;
    documento:string;
    fechaNacimiento: Date;
    correo: String;    
    idsede: String;    
    perfil: String;
    contrasena:string;
    contrasenaAn:string;
    id: number;
    activo: number;
    fechaCreacion: Date;
    UltimaFechaIngreso: Date;
    NombrePerfil: string;
    SedesArray: number;
    EstadoBool: boolean;
    IdPerfil: number;
    ModuloArray:number;
    Permisos:number;
    UsuarioAthenea:string;
    Firma:string;
}